Update example app to use new reusable view
